Added the ability to store and load notes
I made it so we can pass a NoteService objec to every screen. The NoteService allows us to load, add, and save the notes to the file.  Use the notes List in the object to retrieve notes after loading from the file. The notes are also loaded upon opening the app.